Basic functionality to compute additional index scan cost is added.
In this commit following updates are made :

1. To compute additional cost in a scenarios where Index columns are not present
in the predicate,  ComputeAdditionalMissingIndexWeight(), is used.
For all the index columns which are not present in the predicate, this function
computes the missing index column weight. This weight is then used in the CostIndexScan(,)
and CostIndexOnlyScan(,) to compute additional cost.

2. To compute the additional cost when column(s) used in  predicate, is (are) not
present in the index, the 'count of missing predicates columns in index' is used
in  CostIndexScan(,) and CostIndexOnlyScan(,) to compute additional cost.
The missing predicate information is not present in the costing functions.
This information is first added into the different index scans and then retreived
at the time of costing.