Computing NDV for the missing index column
To compute the NDV of the missing index columns at the time of costing,
'Base table stats' were required. These stats hold the information about the
columns of the table. But 'Base table stats' were not generated, it was found that
in CPhysicalScan::ComputeTableStats(), the table columns were not inserted for
Histogram generation in the next step.

This was one of the change made in this commit, along with other changes like
releasing allocated memory for (key,value) pair when insertion is not made in the MAP.