Prevent overwriting of worker in ORCA
In some cases, during optimization of queries with call to an function,
it was found that another instance of ORCA optimzation is launched.

At the start of optimization, a new worker is registered in the
'Worker Pool Manager'. Currently ORCA supports only one worker, if a new
worker is registered on top of an existing worker, then the details of
the old worker will be lost and will cause undefined behaviour.

Thus, through this commit, aim is to stop overwriting of workers.